Moved settings helpers to separate service
