+++ ghost/core/core/server/services/settings-helpers/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./settings-helpers');
