+++ ghost/core/core/server/services/stripe/config.js
@@ -1,5 +1,6 @@
 const logging = require('@tryghost/logging');
 const tpl = require('@tryghost/tpl');
+const SettingsHelpers = require('../settings-helpers');
 
 const messages = {
     remoteWebhooksInDevelopment: 'Cannot use remote webhooks in development. See https://ghost.org/docs/webhooks/#stripe-webhooks for developing with Stripe.'
@@ -41,43 +42,8 @@ module.exports = {
             };
         }
 
-        /**
-         * @param {'direct' | 'connect'} type - The "type" of keys to fetch from settings
-         * @returns {{publicKey: string, secretKey: string} | null}
-         */
-        function getStripeKeys(type) {
-            const secretKey = settings.get(`stripe_${type === 'connect' ? 'connect_' : ''}secret_key`);
-            const publicKey = settings.get(`stripe_${type === 'connect' ? 'connect_' : ''}publishable_key`);
-
-            if (!secretKey || !publicKey) {
-                return null;
-            }
-
-            return {
-                secretKey,
-                publicKey
-            };
-        }
-
-        /**
-         * @returns {{publicKey: string, secretKey: string} | null}
-         */
-        function getActiveStripeKeys() {
-            const stripeDirect = config.get('stripeDirect');
-
-            if (stripeDirect) {
-                return getStripeKeys('direct');
-            }
-
-            const connectKeys = getStripeKeys('connect');
-
-            if (!connectKeys) {
-                return getStripeKeys('direct');
-            }
-
-            return connectKeys;
-        }
-        const keys = getActiveStripeKeys();
+        const helpers = new SettingsHelpers({settingsCache: settings, config, urlUtils});
+        const keys = helpers.getActiveStripeKeys();
         if (!keys) {
             return null;
         }
